requirements.txt
fastapi
uvicorn
pandas
numpy
sqlalchemy
psycopg2-binary
xgboost
scikit-learn
joblib
python-dotenv